+++ src/editor/event-out/editor.js
@@ -10,14 +10,25 @@
      * @param {*} node -
      */
     function onEditPrepare(node) {
-        // initial checkbox states
-        if (!node.passthrough) node.passthrough = false
-        $('#node-input-passthrough')
-            // Initial setting
-            .prop('checked', node.passthrough)
+
+        // Handle v1.0 legacy
+        if (node.passthrough === true) $('#node-input-passthrough').val('input') //node.passthrough = 'input'
+        else if (node.passthrough === false) $('#node-input-passthrough').val('none') //node.passthrough = 'none'
+
+        // Initial state
+        $(`#passthrough-${node.passthrough}`).attr('checked', true)
+
+        // Handle change on radio inputs
+        $('[name="passthrough"]')
             // If the setting changes, change the number of output ports
-            .on('change', function passthroughChange() {
-                node.outputs = this.checked ? 1 : 0
+            .on('change', function passthroughChange(event) {
+                let passthrough = event.target.id.replace('passthrough-', '')
+
+                $('#node-input-passthrough').val(passthrough)
+
+                if ( passthrough === 'none') node.outputs = 0
+                else node.outputs = 1
+
             })
     }
 
@@ -27,17 +38,27 @@
         defaults: {
             name: { value: '' },
             topic: { value: '' },
-            passthrough: { value: false },
+            passthrough: { value: 'none', required:true },
             outputs: { value: 0 },
         },
         align:'right',
         inputs: 1,
         inputLabels: 'Msg with topic property',
         outputs: 0,
-        outputLabels: ['Input msg with node ID added'],
+        outputLabels: function (i) { return this.passthrough === 'return' ? 'msg returned from event-return node' : 'Input msg with node ID added' }, // eslint-disable-line no-unused-vars
         icon: 'link-out.svg',
         paletteLabel: nodeLabel,
         label: function () { return this.name || this.topic || nodeLabel },
+
+        /** Available methods: 
+         * oneditprepare: (function) called when the edit dialog is being built.
+         * oneditsave:   (function) called when the edit dialog is okayed.
+         * oneditcancel: (function) called when the edit dialog is canceled.
+         * oneditdelete: (function) called when the delete button in a configuration node’s edit dialog is pressed.
+         * oneditresize: (function) called when the edit dialog is resized.
+         * onpaletteadd: (function) called when the node type is added to the palette.
+         * onpaletteremove: (function) called when the node type is removed from the palette.
+         */
 
         oneditprepare: function() { onEditPrepare(this) },